[upd] project structure; [add] minimal Vec3 functional
