+++ src/traits.rs
@@ -1,0 +1,4 @@
+pub trait Cross<RHS = Self> {
+    type Output;
+    fn cross(self, rhs: RHS) -> Self::Output;
+}
